+++ shared-components/backend/package/transport/http/middleware/request_id.go
@@ -12,5 +12,6 @@ func RequestIDMiddleware(ctx iris.Context) {
 		requestId = uuid.NewString()
 	}
 	ctx.Values().Set(constants.CTXRequestIdKey, requestId)
+	ctx.Header(constants.CTXRequestIdKey, requestId)
 	ctx.Next()
 }