+++ shared-components/backend/package/utils/utils.go
@@ -1,23 +1,30 @@
 package utils
 
 import (
-	"fmt"
-
+	"tmossDev.github.com/eco-system/shared-components/backend/package/constants"
 	"tmossDev.github.com/eco-system/shared-components/backend/package/logger"
 )
 
-func LogPreparingError(queryName string, err error) {
-	logger.Error(fmt.Sprintf("Error preparing '%s' query: %s", queryName, err.Error()))
+func requestIdOrDefault(requestIds ...string) string {
+	if len(requestIds) > 0 && requestIds[0] != "" {
+		return requestIds[0]
+	}
+
+	return constants.DefaultRequestId
 }
 
-func LogExecutingError(queryName string, err error) {
-	logger.Error(fmt.Sprintf("Error executing '%s' query: %s", queryName, err.Error()))
+func LogPreparingError(queryName string, err error, requestIds ...string) {
+	logger.Errorf(requestIdOrDefault(requestIds...), "Error preparing '%s' query: %s", queryName, err.Error())
 }
 
-func LogBeginingTnxError(queryName string, err error) {
-	logger.Error(fmt.Sprintf("Error begining transaction '%s' query: %s", queryName, err.Error()))
+func LogExecutingError(queryName string, err error, requestIds ...string) {
+	logger.Errorf(requestIdOrDefault(requestIds...), "Error executing '%s' query: %s", queryName, err.Error())
 }
 
-func LogCommitError(queryName string, err error) {
-	logger.Error(fmt.Sprintf("Error commiting transaction '%s' query: %s", queryName, err.Error()))
+func LogBeginingTnxError(queryName string, err error, requestIds ...string) {
+	logger.Errorf(requestIdOrDefault(requestIds...), "Error begining transaction '%s' query: %s", queryName, err.Error())
+}
+
+func LogCommitError(queryName string, err error, requestIds ...string) {
+	logger.Errorf(requestIdOrDefault(requestIds...), "Error commiting transaction '%s' query: %s", queryName, err.Error())
 }