+++ user-management/backend/package/user/repository/user.go
@@ -6,7 +6,7 @@ import (
 
 type UserRepository interface {
 	GetByID(userId uint64) (*model.UserResponse, error)
-	GetByEmail(email string) (*model.UserResponse, error)
+	GetByEmail(requestId string, email string) (*model.UserResponse, error)
 	RegisterUser(firstName string, lastName string, email string, password string, roleId uint64) (*model.UserResponse, error)
 	Update(userId uint64, firstName string, lastName string, updatingUserId uint64) (*model.UserResponse, error)
 	ResetPassword(userId uint64, newPassword string, updatingUserId uint64) (*model.UserResponse, error)